Made separate files to handle each table for postgres.
