+++ postgres/psPool.js
@@ -1,0 +1,13 @@
+const path = require('path')
+require('dotenv').config({ path: path.resolve(__dirname, '../.env') });
+const { Pool } = require('pg');
+
+
+const pool = new Pool({
+  connectionString: process.env.DATABASE_URL,
+  //ssl: { rejectUnauthorized: false }
+});
+
+module.exports = {
+  pool: pool
+};
